[refactor] 가독성 및 책임 분리를 위해 service layer에 exception layer 추가
