+++ src/main/java/junseok/snr/wallet/application/service/EtherWalletService.java
@@ -1,6 +1,8 @@
 package junseok.snr.wallet.application.service;
 
 import io.micrometer.common.util.StringUtils;
+import junseok.snr.wallet.application.service.exception.ExceptionCode;
+import junseok.snr.wallet.application.service.exception.WalletException;
 import junseok.snr.wallet.infrastructure.common.Web3jUtils;
 import junseok.snr.wallet.api.dto.CreateWalletRequestDto;
 import junseok.snr.wallet.domain.model.Wallet;
