+++ src/main/java/junseok/snr/wallet/application/service/exception/WalletException.java
@@ -1,4 +1,4 @@
-package junseok.snr.wallet.application.service;
+package junseok.snr.wallet.application.service.exception;
 
 import lombok.Getter;
